fix main card spacing
