+++ app/stats/page-client.tsx
@@ -5,6 +5,8 @@ import { TimeToFillCard } from "@/app/stats/charts/time-to-fill-card"
 import { TvlChart } from "@/app/stats/charts/tvl-chart"
 import { VolumeChart } from "@/app/stats/charts/volume-chart"
 
+import { Button } from "@/components/ui/button"
+
 export function PageClient() {
   return (
     <main className="container mb-8 mt-12 flex flex-col gap-12 px-4 lg:px-8">
@@ -13,7 +15,27 @@ export function PageClient() {
           <h1 className="mb-4 mt-6 font-serif text-3xl font-bold tracking-tighter lg:tracking-normal">
             Time to Fill
           </h1>
-          <TimeToFillCard />
+          <div className="grid grid-cols-[3fr_1fr_1fr] gap-4">
+            <TimeToFillCard />
+            <div className="grid place-items-center border p-6">
+              <div className="text-4xl">Zero</div>
+              <div className="flex flex-col items-center gap-1 text-sm text-muted-foreground">
+                <div className="transition-colors hover:text-primary">
+                  Price Impact
+                </div>
+                <div className="transition-colors hover:text-primary">
+                  Copy Trading
+                </div>
+                <div className="transition-colors hover:text-primary">
+                  Slippage
+                </div>
+                <div className="transition-colors hover:text-primary">MEV</div>
+              </div>
+            </div>
+            <div className="grid place-items-center border p-6">
+              <Button variant="ghost">Place Order</Button>
+            </div>
+          </div>
         </div>
       </div>
